Audio-driven frame timing for accurate ~59.73 Hz
- APU: Use fractional sample counter (48000/4194304) for exact 48kHz
- AudioBuffer: Add throttling when 75%+ full to sync to real-time
- Window: Remove SDL PRESENTVSYNC (now audio-driven)
- Docs: Add Frame Timing section, update test counts

Internal hardware timing unchanged, verified 31/35 Mooneye passing.

+++ src/apu/APU.hpp
@@ -158,7 +158,7 @@ private:
     float left_sample;
     float right_sample;
     bool sample_ready;
-    uint16_t sample_counter;        // For downsampling to target rate
+    uint32_t sample_counter;        // Fractional counter for accurate 48kHz downsampling
     AudioBuffer* audio_buffer = nullptr;  // External buffer for SDL audio
     
     // === Internal Operations (directly expose the channel logic) ===
